Move String Literal to literals file
